Add v0.1 of split panes
Add a new split pane button in the top right that when clicked,
enables a new right split and puts "focus" in that split.
Any definitions that are opened are opened in the focused pane.

Toggling the right pane off again, keeps the open definitions, allowing
them to be shown when the pane is enabled once more.

This does support rudimentary resizing of the definition cards, but
doesn't support proper scrolling of code as panes are resized.