Add support for switching pane focus with keyboard
Using w+arrow-right or w+arrow-left (or h/l) will move the focus from
one pane to the other (any definitions opened will open in the focused
pane). Currently this is indicated by the "non-focused" pane having a
lower opacity. More testing is required to determine if that is a
sufficient affordance.