Migrate from Tauri to Electron
While truly a cool project, Tauri has proven to not be the right move
for Unison at this time. The challenge revolve almost entirely around
one of its major appeals: not bundling a browser rendering engine with
the application. This means we can't consistently rely on web features
across all platforms or older versions of platforms (testing on an old
MacOS is increasingly difficult). Additionally the performance of the
webview on Linux is a concern as we have users spread across all
platforms.

This commit fully migrates the app to run on Electron, but is missing
work needed for distribution.